style: move all wire set into wire.go
