+++ cmd/cgtproxy/cmd/wire.go
@@ -25,3 +25,25 @@ func injectedLastingCGTProxy(
 ) {
 	panic(wire.Build(lastingConnectorSet))
 }
+
+var set = wire.NewSet(
+	provideBypass,
+	provideCGTProxy,
+	provideCGroupEventChan,
+	provideCgrougMontior,
+	provideCgroupRoot,
+	provideNFTManager,
+	provideNetlinkConnector,
+	provideRuleManager,
+)
+
+var lastingConnectorSet = wire.NewSet(
+	provideBypass,
+	provideCGTProxy,
+	provideCGroupEventChan,
+	provideCgrougMontior,
+	provideCgroupRoot,
+	provideLastringNetlinkConnector,
+	provideNFTManager,
+	provideRuleManager,
+)
